pyproject.toml
[tool.poetry]
name = "dvc-parallel-runner-example"
version = "0.1.0"
description = "Repo demonstrating usage of Neuromancer tool for parallel execution of DVC projects on k8s clusters"
authors = ["Sami Jawhar <sami@thecybermonk.com>"]
readme = "README.md"
packages = [{ include = "pipelines" }]

[tool.poetry.dependencies]
python = "^3.9"

click = "^8.1.3"
dvc = { version = "^2.37.0", extras = ["s3"] }
matplotlib = "^3.6.2"
numpy = "^1.23.5"
pandas = "^1.5.2"
pyyaml = "^6.0"
scikit-learn = "^1.2.0"
scipy = "^1.9.3"

[tool.poetry.group.gpu]
optional = true
[tool.poetry.group.gpu.dependencies]
cupy-cuda11x = "^11.4.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
debugpy = "^1.6.4"
pytest = "^7.2.0"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py39"]

[tool.isort]
known_local_folder = "pipelines"
profile = "black"
